Gather ProjectManager items component in dedicated file
