+++ newIDE/app/src/ProjectManager/ProjectManagerItems.js
@@ -1,0 +1,304 @@
+// @flow
+import * as React from 'react';
+import { t } from '@lingui/macro';
+import { I18n } from '@lingui/react';
+import { type I18n as I18nType } from '@lingui/core';
+import WarningIcon from '@material-ui/icons/Warning';
+
+import { type MenuItemTemplate } from '../UI/Menu/Menu.flow';
+import { IconContainer } from '../UI/IconContainer';
+import { ListItem } from '../UI/List';
+import ThemeConsumer from '../UI/Theme/ThemeConsumer';
+import TextField, {
+  noMarginTextFieldInListItemTopOffset,
+} from '../UI/TextField';
+import { shouldValidate } from '../UI/KeyboardShortcuts/InteractionKeys';
+import { textEllispsisStyle } from '../UI/TextEllipsis';
+
+import { ExtensionStoreContext } from '../AssetStore/ExtensionStore/ExtensionStoreContext';
+import { type ExtensionShortHeader } from '../Utils/GDevelopServices/Extension';
+
+const styles = {
+  noIndentNestedList: {
+    padding: 0,
+  },
+  itemTextField: {
+    top: noMarginTextFieldInListItemTopOffset,
+  },
+};
+
+type ProjectStructureItemProps = {|
+  autoGenerateNestedIndicator?: boolean,
+  initiallyOpen?: boolean,
+  leftIcon?: React$Element<any>,
+  indentNestedItems?: boolean,
+  renderNestedItems: () => Array<React$Element<any> | null>,
+  primaryText: React.Node,
+  error?: ?Error,
+  onRefresh?: () => void,
+  open?: boolean,
+|};
+
+export const ProjectStructureItem = (props: ProjectStructureItemProps) => (
+  <ThemeConsumer>
+    {muiTheme => {
+      const {
+        error,
+        leftIcon,
+        onRefresh,
+        indentNestedItems,
+        ...otherProps
+      } = props;
+      return (
+        <ListItem
+          style={{
+            backgroundColor: muiTheme.listItem.groupBackgroundColor,
+            borderBottom: `1px solid ${muiTheme.listItem.separatorColor}`,
+          }}
+          nestedListStyle={
+            indentNestedItems ? undefined : styles.noIndentNestedList
+          }
+          {...otherProps}
+          leftIcon={error ? <WarningIcon /> : leftIcon}
+          displayReloadButton={!!error}
+          onReload={onRefresh}
+          reloadButtonTooltip={`An error has occured in functions. Click to reload them.`}
+        />
+      );
+    }}
+  </ThemeConsumer>
+);
+
+type ItemProps = {|
+  primaryText: string,
+  editingName: boolean,
+  leftIcon?: ?React.Node,
+  onEdit: () => void,
+  onDelete: () => void,
+  addLabel: string,
+  onAdd: () => void,
+  onRename: string => void,
+  onEditName: () => void,
+  onCopy: () => void,
+  onCut: () => void,
+  onPaste: () => void,
+  onDuplicate: () => void,
+  canPaste: () => boolean,
+  canMoveUp: boolean,
+  onMoveUp: () => void,
+  canMoveDown: boolean,
+  onMoveDown: () => void,
+  buildExtraMenuTemplate?: (i18n: I18nType) => Array<MenuItemTemplate>,
+  style?: ?Object,
+|};
+
+export const Item = ({
+  primaryText,
+  editingName,
+  leftIcon,
+  onEdit,
+  onDelete,
+  addLabel,
+  onAdd,
+  onRename,
+  onEditName,
+  onCopy,
+  onCut,
+  onPaste,
+  onDuplicate,
+  canPaste,
+  canMoveUp,
+  onMoveUp,
+  canMoveDown,
+  onMoveDown,
+  buildExtraMenuTemplate,
+  style,
+}: ItemProps) => {
+  const textField = React.useRef<?TextField>(null);
+
+  React.useEffect(
+    () => {
+      if (editingName)
+        setTimeout(() => {
+          if (textField.current) textField.current.focus();
+        }, 100);
+    },
+    [editingName]
+  );
+
+  const label = editingName ? (
+    <TextField
+      id="rename-item-field"
+      margin="none"
+      ref={textField}
+      defaultValue={primaryText}
+      onBlur={e => onRename(e.currentTarget.value)}
+      onKeyPress={event => {
+        if (shouldValidate(event)) {
+          if (textField.current) textField.current.blur();
+        }
+      }}
+      fullWidth
+      style={styles.itemTextField}
+    />
+  ) : (
+    <div style={textEllispsisStyle} title={primaryText}>
+      {primaryText}
+    </div>
+  );
+
+  return (
+    <ThemeConsumer>
+      {muiTheme => (
+        <I18n>
+          {({ i18n }) => (
+            <ListItem
+              style={{
+                borderBottom: `1px solid ${muiTheme.listItem.separatorColor}`,
+                ...style,
+              }}
+              primaryText={label}
+              leftIcon={leftIcon}
+              displayMenuButton
+              buildMenuTemplate={(i18n: I18nType) => [
+                {
+                  label: i18n._(t`Edit`),
+                  click: () => onEdit(),
+                },
+                ...(buildExtraMenuTemplate ? buildExtraMenuTemplate(i18n) : []),
+                { type: 'separator' },
+                {
+                  label: i18n._(t`Rename`),
+                  click: () => onEditName(),
+                },
+                {
+                  label: i18n._(t`Delete`),
+                  click: () => onDelete(),
+                },
+                {
+                  label: addLabel,
+                  visible: !!onAdd,
+                  click: () => onAdd(),
+                },
+                { type: 'separator' },
+                {
+                  label: i18n._(t`Copy`),
+                  click: () => onCopy(),
+                },
+                {
+                  label: i18n._(t`Cut`),
+                  click: () => onCut(),
+                },
+                {
+                  label: i18n._(t`Paste`),
+                  enabled: canPaste(),
+                  click: () => onPaste(),
+                },
+                {
+                  label: i18n._(t`Duplicate`),
+                  click: () => onDuplicate(),
+                },
+                { type: 'separator' },
+                {
+                  label: i18n._(t`Move up`),
+                  enabled: canMoveUp,
+                  click: () => onMoveUp(),
+                },
+                {
+                  label: i18n._(t`Move down`),
+                  enabled: canMoveDown,
+                  click: () => onMoveDown(),
+                },
+              ]}
+              onClick={() => {
+                // It's essential to discard clicks when editing the name,
+                // to avoid weird opening of an editor (accompanied with a
+                // closing of the project manager) when clicking on the text
+                // field.
+                if (!editingName) onEdit();
+              }}
+            />
+          )}
+        </I18n>
+      )}
+    </ThemeConsumer>
+  );
+};
+
+type EventFunctionExtensionItemProps = {|
+  eventsFunctionsExtension: gdEventsFunctionsExtension,
+  onEdit: ({ [string]: ExtensionShortHeader }) => void,
+  onRename: string => void,
+  onEditName: () => void,
+  isEditingName: boolean,
+  onDelete: () => void,
+  onAdd: () => void,
+  onCopy: () => void,
+  onCut: () => void,
+  onPaste: () => void,
+  onDuplicate: () => void,
+  canPaste: () => boolean,
+  canMoveUp: boolean,
+  onMoveUp: () => void,
+  canMoveDown: boolean,
+  onMoveDown: () => void,
+|};
+
+export const EventFunctionExtensionItem = ({
+  eventsFunctionsExtension,
+  onEdit,
+  onRename,
+  onEditName,
+  isEditingName,
+  onDelete,
+  onAdd,
+  onCopy,
+  onCut,
+  onPaste,
+  onDuplicate,
+  canPaste,
+  canMoveUp,
+  onMoveUp,
+  canMoveDown,
+  onMoveDown,
+}: EventFunctionExtensionItemProps) => {
+  const name = eventsFunctionsExtension.getName();
+  const iconUrl = eventsFunctionsExtension.getIconUrl();
+
+  const { extensionShortHeadersByName } = React.useContext(
+    ExtensionStoreContext
+  );
+
+  return (
+    <Item
+      leftIcon={
+        iconUrl ? (
+          <IconContainer
+            size={24}
+            alt={eventsFunctionsExtension.getFullName()}
+            src={iconUrl}
+          />
+        ) : null
+      }
+      primaryText={name}
+      editingName={isEditingName}
+      onEdit={() => onEdit(extensionShortHeadersByName)}
+      onDelete={onDelete}
+      addLabel={'Add a New Extension'} // TODO: Translate this
+      onAdd={onAdd}
+      onRename={newName => {
+        onRename(newName);
+      }}
+      onEditName={onEditName}
+      onCopy={onCopy}
+      onCut={onCut}
+      onPaste={onPaste}
+      onDuplicate={onDuplicate}
+      canPaste={canPaste}
+      canMoveUp={canMoveUp}
+      onMoveUp={onMoveUp}
+      canMoveDown={canMoveDown}
+      onMoveDown={onMoveDown}
+    />
+  );
+};
